+++ modules/sdk-coin-eth2/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.27](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-eth2@3.0.26...@bitgo/sdk-coin-eth2@3.0.27) (2024-07-30)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-eth2
+
 ## [3.0.26](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-eth2@3.0.25...@bitgo/sdk-coin-eth2@3.0.26) (2024-07-24)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-eth2